Cherry pick small fixes out of spike source array branch * Removed bad pasting * Setting of log level from make command line * missed bit of review * WHOOPS - output core shouldn't read from the same buffer input is writing to! * Config option to determine whether stop signal gets sent i.e. for prodding in ybug afterwards

+++ simulator.py
@@ -80,7 +80,7 @@ class State(common.control.BaseState):
         self.segment_counter += 1
 
     def stop(self):
-        if self.machine_controller is not None:
+        if self.machine_controller is not None and self.stop_on_spinnaker:
             logger.info("Stopping SpiNNaker application")
             self.machine_controller.send_signal("stop")
